qa_evaluation_system/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.62.0

# NLP and ML
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
nltk>=3.8.0
spacy>=3.5.0
textstat>=0.7.3

# LLM APIs
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Metrics
rouge-score>=0.1.2
sacrebleu>=2.3.0
bert-score>=0.3.13

# Chinese NLP support
jieba>=0.42.1
zhon>=1.1.5
pypinyin>=0.49.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
jsonschema>=4.0.0

# Web framework (optional, for API service)
fastapi>=0.100.0
uvicorn>=0.23.0

# Data processing
openpyxl>=3.0.0
jsonlines>=3.1.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0